+++ examples/live_open_orders_request.rs
@@ -3,7 +3,7 @@ use kraken_async_rs::clients::http_response_types::ResultErrorResponse;
 use kraken_async_rs::clients::kraken_client::KrakenClient;
 use kraken_async_rs::crypto::nonce_provider::{IncreasingNonceProvider, NonceProvider};
 use kraken_async_rs::request_types::OpenOrdersRequest;
-use kraken_async_rs::secrets::secrets_provider::EnvSecretsProvider;
+use kraken_async_rs::secrets::secrets_provider::{EnvSecretsProvider, SecretsProvider};
 use std::sync::Arc;
 use tokio::sync::Mutex;
 
@@ -12,7 +12,10 @@ use tokio::sync::Mutex;
 async fn main() {
     // note that this will fail if you don't have your key and secret set to these env vars
     // eg `export KRAKEN_KEY="YOUR-API-KEY"`, ...
-    let secrets_provider = Box::new(EnvSecretsProvider::new("KRAKEN_KEY", "KRAKEN_SECRET"));
+    let secrets_provider: Box<Arc<Mutex<dyn SecretsProvider>>> = Box::new(Arc::new(Mutex::new(
+        EnvSecretsProvider::new("KRAKEN_KEY", "KRAKEN_SECRET"),
+    )));
+
     let nonce_provider: Box<Arc<Mutex<dyn NonceProvider>>> =
         Box::new(Arc::new(Mutex::new(IncreasingNonceProvider::new())));