+++ examples/live_query_orders_info.rs
@@ -2,7 +2,7 @@ use kraken_async_rs::clients::core_kraken_client::CoreKrakenClient;
 use kraken_async_rs::clients::kraken_client::KrakenClient;
 use kraken_async_rs::crypto::nonce_provider::{IncreasingNonceProvider, NonceProvider};
 use kraken_async_rs::request_types::OrderRequest;
-use kraken_async_rs::secrets::secrets_provider::EnvSecretsProvider;
+use kraken_async_rs::secrets::secrets_provider::{EnvSecretsProvider, SecretsProvider};
 use std::fs::File;
 use std::sync::Arc;
 use tokio::sync::Mutex;
@@ -15,7 +15,10 @@ async fn main() {
     set_up_logging("tests.log");
     // note that this will fail if you don't have your key and secret set to these env vars
     // eg `export KRAKEN_KEY="YOUR-API-KEY"`, ...
-    let secrets_provider = Box::new(EnvSecretsProvider::new("KRAKEN_KEY", "KRAKEN_SECRET"));
+    let secrets_provider: Box<Arc<Mutex<dyn SecretsProvider>>> = Box::new(Arc::new(Mutex::new(
+        EnvSecretsProvider::new("KRAKEN_KEY", "KRAKEN_SECRET"),
+    )));
+
     let nonce_provider: Box<Arc<Mutex<dyn NonceProvider>>> =
         Box::new(Arc::new(Mutex::new(IncreasingNonceProvider::new())));