+++ examples/live_wss_trading.rs
@@ -3,7 +3,7 @@ use kraken_async_rs::clients::kraken_client::KrakenClient;
 use kraken_async_rs::crypto::nonce_provider::{IncreasingNonceProvider, NonceProvider};
 use kraken_async_rs::request_types::TimeInForce;
 use kraken_async_rs::response_types::{BuySell, OrderFlag, OrderType};
-use kraken_async_rs::secrets::secrets_provider::EnvSecretsProvider;
+use kraken_async_rs::secrets::secrets_provider::{EnvSecretsProvider, SecretsProvider};
 use kraken_async_rs::wss::kraken_wss_client::{KrakenMessageStream, KrakenWSSClient};
 use kraken_async_rs::wss::private::messages::PrivateMessage;
 use kraken_async_rs::wss::private::trading_messages::AddOrderRequest;
@@ -24,7 +24,9 @@ use tokio_stream::StreamExt;
 /// which goes in any request made via the websocket connection.
 #[tokio::main]
 async fn main() {
-    let secrets_provider = Box::new(EnvSecretsProvider::new("KRAKEN_KEY", "KRAKEN_SECRET"));
+    let secrets_provider: Box<Arc<Mutex<dyn SecretsProvider>>> = Box::new(Arc::new(Mutex::new(
+        EnvSecretsProvider::new("KRAKEN_KEY", "KRAKEN_SECRET"),
+    )));
     let nonce_provider: Box<Arc<Mutex<dyn NonceProvider>>> =
         Box::new(Arc::new(Mutex::new(IncreasingNonceProvider::new())));
     let mut kraken_client = CoreKrakenClient::new(secrets_provider, nonce_provider);