+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### In Progress
 
+- Change `KrakenClient` signature to require `secrets_provider: Box<Arc<Mutex<dyn SecretsProvider>>>,` to enable `Clone`
+  on `CoreKrakenClient`
 - Fix `BalancesSubscription` sending null for snapshot by default (skip_serializing_none)
 - Fix `WSSMessage` by adding WssMessage(ErrorResponse { .. }) variant
 - Change `checksum` fields for L2 and L3 books to `u32` to match crc32fast crate impl